policy/lab/service/requirements.txt
Babel==2.3.4
PyYAML==3.12
Pygments==2.2.0
WebOb==1.7.1
appdirs==1.4.3
argparse==1.2.1
certifi==2017.4.17
chardet==3.0.4
cliff==2.7.0
cmd2==0.7.2
debtcollector==1.15.0
deprecation==1.0.1
falcon==1.2.0
funcsigs==1.0.2
functools32==3.2.3-2
httpie==0.9.9
idna==2.5
iso8601==0.1.11
jsonpatch==1.15
jsonpointer==1.10
jsonschema==2.6.0
keystoneauth1==2.21.0
keystonemiddleware==4.14.0
monotonic==1.3
msgpack-python==0.4.8
netaddr==0.7.19
netifaces==0.10.6
openstacksdk==0.9.17
os-client-config==1.27.0
os-http==0.1.0
osc-lib==1.6.0
oslo.config==4.5.0
oslo.context==2.13.0
oslo.i18n==3.15.3
oslo.log==3.20.1
oslo.serialization==2.18.0
oslo.utils==3.26.0
pbr==3.0.1
positional==1.1.1
prettytable==0.7.2
pycadf==2.5.0
pyinotify==0.9.6
pyparsing==2.2.0
python-cinderclient==2.2.0
python-dateutil==2.6.0
python-glanceclient==2.7.0
python-keystoneclient==3.11.0
python-mimeparse==1.6.0
python-novaclient==9.0.0
python-openstackclient==3.11.0
pytz==2017.2
requests==2.18.1
requestsexceptions==1.2.0
rfc3986==1.0.0
simplejson==3.10.0
six==1.10.0
stevedore==1.23.0
ujson==1.35
unicodecsv==0.14.1
urllib3==1.21.1
warlock==1.2.0
wrapt==1.10.10
wsgiref==0.1.2
